# doc/api.yaml
openapi: 3.0.3
info:
  title: WasaText
  description: | 
    This is the yaml documentation for the WasaText Api
  version: 1.0.0
paths:
...
  /conversations/start-conversation:
    post:
      tags:
        - conversations
      summary: Start a new private conversation (1:1)
      description: >
        Starts a new conversation with another user and returns the id of the conversation.
      operationId: createPrivateConversation
      security:
        - bearerAuth: []
      requestBody:
        description: Conversation details
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
              properties:
                username:
                  type: string
                  example: "hanniPham"
      responses:
        '200':
          description: Existing conversation found
          content:
            application/json:
              schema:
                type: object
                properties:
                  conversation_id:
                    type: string
                    example: "1"
        '201':
          description: New conversation created
          content:
            application/json:
              schema:
                type: object
                properties:
                  conversation_id:
                    type: string
                    example: "1"
        '400':
          description: Invalid request
        '404':
          description: User not found
...
  /conversations/delete/{conversation_id}:
    delete:
      tags:
        - conversations
      summary: Delete a conversation
      operationId: deleteConversation
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/conversation_id'
      responses:
        '204':
          description: Conversation deleted successfully
        '404':
          description: Conversation not found
        '403':
          description: Forbidden
...
components:
  parameters:
...
    conversation_id:
      schema:
        type: string
      name: conversation_id
      in: path
      required: true
      description: The conversation ID
      allowEmptyValue: false
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...